requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
neo4j
pydantic-settings
pytest
PyYAML
jinja2